refactor: 인라인 코드로 작성된 props를 hook으로 분리
